Bug fix
DSP (Display sprite) now has a proper `void` return instead of the `bool`
because it didn't return anything (so I think that means it defaults to false
or something) which screwed up a large amount of the pixel by pixel collisions
and stuff. Everything works now, except for the Simon Says one. The BRIX and
TETRIS worked, which was/is a relief.