Added a (somewhat) finished assembler
Assembler follows a slight modified version of the mnemonics described on
[Cowgod's Chip-8 Technical Reference](http://devernay.free.fr/hacks/chip8/C8TECH10.HTM)
(To the best of my memory), only 4 mnemonics have been altered (the last 4).
Nonetheless, the mnemonics are all there for you in the docs/assembler.md file
for you all to look at (you know, if you want to). The disassembler is modified
so that it can work with the assembler such that:

Let D be the function to disassemble chip8 binary.
Let A be the function to assemble chip8 source.
Let F be the original chip8 binary.

A(D(F)) = F

Unfortunately, it does not work with extremely large source files. You would
get an idea of what "very large" means by disassembling the BLINKY binary
and plugging the whole thing into the assembler.

Will prepare for merge with master branch (whatever that means).